requirements.txt
asgiref==3.7.1
Django==4.2.1
mysql-connector-python==8.0.33
protobuf==3.20.3
sqlparse==0.4.4
